set to true by default if oper_status is not set.
