+++ src/main/java/org/o3project/odenos/component/linklayerizer/LinkLayerizerOnFlow.java
@@ -20,8 +20,8 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.o3project.odenos.core.component.ConversionTable;
 import org.o3project.odenos.core.component.Logic;
-import org.o3project.odenos.core.component.Logic.AttrElements;
-import org.o3project.odenos.core.component.Logic.NetworkElements;
+import org.o3project.odenos.core.component.Component.AttrElements;
+import org.o3project.odenos.core.component.Component.NetworkElements;
 import org.o3project.odenos.core.component.NetworkInterface;
 import org.o3project.odenos.core.component.network.flow.Flow;
 import org.o3project.odenos.core.component.network.flow.FlowObject.FlowStatus;
