+++ src/test/java/org/o3project/odenos/component/linklayerizer/LinkLayerizerOnFlowTest.java
@@ -34,8 +34,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.o3project.odenos.core.component.ConversionTable;
-import org.o3project.odenos.core.component.Logic.AttrElements;
-import org.o3project.odenos.core.component.Logic.NetworkElements;
+import org.o3project.odenos.core.component.Component.AttrElements;
+import org.o3project.odenos.core.component.Component.NetworkElements;
 import org.o3project.odenos.core.component.NetworkInterface;
 import org.o3project.odenos.core.component.network.flow.Flow;
 import org.o3project.odenos.core.component.network.flow.FlowObject.FlowStatus;
